Start of work on f.s

+++ f.h
@@ -1,7 +1,7 @@
 #ifndef F_H_
 #define F_H_
 
-void drawQuadratic(unsigned char * pPixelBuffer, int width, int height, int a, int b, int c, int S);
+void drawQuadratic(unsigned char * pPixelBuffer, int width, int height, double a, double b, double c, double S);
 
 #endif // F_H_
